fix function and add id automatic

+++ Library/customer/views.py
@@ -9,12 +9,12 @@ customer_table = Blueprint('Customers' ,__name__,template_folder="templates")
 @customer_table.route('/addCustomer/', methods=['POST', 'GET'])
 def add_customer():
     if request.method == 'POST':
-        if int(request.form['ID']) >= 1 and 100 >= int(request.form['Age']) >= 1 :
-            NewCustomer = Customers(CusomerID=request.form['ID'], Name=request.form['Name'],
+        if 100 >= int(request.form['Age']) >= 1:
+            NewCustomer = Customers(Name=request.form['Name'],
                                     City=request.form['City'], Age=request.form['Age'])
             db.session.add(NewCustomer)
             db.session.commit()
-            return render_template('AddCust.html')
+            return render_template('ShowAllCust.html',print_all_customer = Customers.query.all())
         else:
             return render_template('Eror.html')
     return render_template('AddCust.html')
@@ -35,10 +35,7 @@ def print_all_customer(customername=''):
 @customer_table.route("/deleteCustomer/<id>", methods=['DELETE', 'GET'])
 def delete_customer(id=0):
         customerID=Customers.query.get(int(id))
-        custInLoan=Loans.query.all()
-        for id_cust in custInLoan:
-            return render_template('Eror.html')
-        else: 
+        if customerID:
             db.session.delete(customerID)
             db.session.commit()
-        return render_template('ShowAllCust.html', PrintAllCustomer = Customers.query.all())
+            return render_template('ShowAllCust.html',print_all_customer = Customers.query.all())
